src/main/java/com/bookcatalog/validation/PasswordMatchesValidator.java
package com.bookcatalog.validation;

import com.bookcatalog.model.UserRegistrationDto;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;

public class PasswordMatchesValidator
        implements ConstraintValidator<PasswordMatches, UserRegistrationDto> {

    @Override
    public void initialize(PasswordMatches constraintAnnotation) {
    }

    @Override
    public boolean isValid(UserRegistrationDto userRegistrationDto, ConstraintValidatorContext context) {
        return userRegistrationDto.getPassword().equals(userRegistrationDto.getConfirmPassword());
    }
}
